Модуляция. Правки.
