+++ js/constants.js
@@ -1,0 +1,5 @@
+const MAX_STRING_LENGTH = 140;
+
+const MAX_ID = 25;
+
+export {MAX_STRING_LENGTH, MAX_ID};
