+++ js/main.js
@@ -1,2 +1,3 @@
 import {photos} from './data.js';
+
 photos();
